notifier: move PostableMimirReceiverToPostableGrafanaReceiver to legacy_storage/v1 package
